change big_num to 32 bit array

+++ Implementierung/libs/big_num/big_num.h
@@ -1,8 +1,9 @@
 
+#include <cstdint>
 #include <stdint.h>
 #include <stdlib.h>
 
 struct bignum {
     size_t size;
-    uint64_t *digits;
+    uint32_t *digits;
 };
